Move wallet status
