+++ wallets/card.js
@@ -3,12 +3,12 @@ import styles from '@/styles/wallet.module.css'
 import Plug from '@/svgs/plug.svg'
 import Gear from '@/svgs/settings-5-fill.svg'
 import Link from 'next/link'
-import { Status, isConfigured } from '@/wallets/common'
+import { isConfigured } from '@/wallets/common'
 import DraggableIcon from '@/svgs/draggable.svg'
 import RecvIcon from '@/svgs/arrow-left-down-line.svg'
 import SendIcon from '@/svgs/arrow-right-up-line.svg'
 import { useWalletImage } from '@/wallets/image'
-import { useWalletStatus, statusToClass } from '@/wallets/status'
+import { useWalletStatus, statusToClass, Status } from '@/wallets/status'
 import { useWalletSupport } from '@/wallets/support'
 
 export default function WalletCard ({ wallet, draggable, onDragStart, onDragEnter, onDragEnd, onTouchStart, sourceIndex, targetIndex, index }) {
