+++ wallets/common.js
@@ -1,12 +1,5 @@
 import walletDefs from '@/wallets/client'
 
-export const Status = {
-  Enabled: 'Enabled',
-  Disabled: 'Disabled',
-  Error: 'Error',
-  Warning: 'Warning'
-}
-
 export function getWalletByName (name) {
   return walletDefs.find(def => def.name === name)
 }
